// src/types/domain/patientClassification.ts
export enum Specialty {
  MEDICINA = 'Med Interna',
  CIRUGIA = 'Cirugía',
  TRAUMATOLOGIA = 'Traumatología',
  GINECOBSTETRICIA = 'Ginecobstetricia',
  PSIQUIATRIA = 'Psiquiatría',
  PEDIATRIA = 'Pediatría',
  ODONTOLOGIA = 'Odontología',
  OTRO = 'Otro',
  EMPTY = '',
}

export enum PatientStatus {
  GRAVE = 'Grave',
  DE_CUIDADO = 'De cuidado',
  ESTABLE = 'Estable',
  EMPTY = '',
}